requirements.txt
streamlit>=1.35
SQLAlchemy>=2.0
python-dotenv>=1.0
pandas>=2.2
openpyxl>=3.1
tzdata>=2024.1
pyngrok>=7.0